// src/app/demo/models/modelsgeneral/subcategoriaviewmodel.ts
export class subcategoria {
    suca_Id?: number;
    suca_Descripcion?: string;
    cate_Id?: string;
    cate_Descripcion?: string;
    usua_Creacion?: number;
    usua_Modificacion?: number;
    usuaCreacion?: string;
    usuaModificacion?: string;
    suca_FechaCreacion?: string;
    suca_FechaModificacion?: string;
    codigo?: number;
}

export class drop {
    cate_Id?: number;
    cate_Descripcion?: string;
}